dashboard-detection/requirements.txt
streamlit
google-cloud-storage
streamlit-autorefresh
matplotlib
pandas
psycopg2-binary
